Reorganiza código em pacote lia
