+++ lia/dialog_manager.py
@@ -1,4 +1,4 @@
-from config import config
+from lia.config import config
 
 
 class DialogManager:
