+++ tests/test_plugin_loader.py
@@ -16,7 +16,7 @@ ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if ROOT_DIR not in sys.path:
     sys.path.insert(0, ROOT_DIR)
 
-from plugins.loader import load_plugins
+from lia.plugins.loader import load_plugins
 
 
 def test_load_plugins_filters_invalid_modules():
